vision/CelebaLandmarks_FlyAI/vgg19/requirements.txt
scikit-image
flyai
Pillow
pandas
tensorflow
torchvision
numpy
torch
Requests
utils
fastai
matplotlib
